refactor(core): add /edit and /add routes for address forms
