+++ .changeset/lazy-actors-drop.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": minor
+---
+
+Move address forms in account to their own /add and /edit pages.
