+++ core/app/[locale]/(default)/account/(tabs)/addresses/add/_actions/add-address.ts
@@ -6,18 +6,17 @@ import { getTranslations } from 'next-intl/server';
 import { getSessionCustomerId } from '~/auth';
 import { client } from '~/client';
 import { graphql, VariablesOf } from '~/client/graphql';
-import { parseAccountFormData } from '~/components/form-fields';
+import { parseAccountFormData } from '~/components/form-fields/shared/parse-fields';
 
-const UpdateCustomerAddressMutation = graphql(`
-  mutation UpdateCustomerAddressMutation(
-    $input: UpdateCustomerAddressInput!
+const AddCustomerAddressMutation = graphql(`
+  mutation AddCustomerAddressMutation(
+    $input: AddCustomerAddressInput!
     $reCaptchaV2: ReCaptchaV2Input
   ) {
     customer {
-      updateCustomerAddress(input: $input, reCaptchaV2: $reCaptchaV2) {
+      addCustomerAddress(input: $input, reCaptchaV2: $reCaptchaV2) {
         errors {
-          __typename
-          ... on CustomerAddressUpdateError {
+          ... on CustomerAddressCreationError {
             message
           }
           ... on CustomerNotLoggedInError {
@@ -38,12 +37,10 @@ const UpdateCustomerAddressMutation = graphql(`
   }
 `);
 
-type Variables = VariablesOf<typeof UpdateCustomerAddressMutation>;
-export type UpdateCustomerAddressInput = Variables['input'];
+type Variables = VariablesOf<typeof AddCustomerAddressMutation>;
+type AddCustomerAddressInput = Variables['input'];
 
-const isUpdateCustomerAddressInput = (
-  data: unknown,
-): data is UpdateCustomerAddressInput['data'] => {
+const isAddCustomerAddressInput = (data: unknown): data is AddCustomerAddressInput => {
   if (typeof data === 'object' && data !== null && 'address1' in data) {
     return true;
   }
@@ -51,23 +48,21 @@ const isUpdateCustomerAddressInput = (
   return false;
 };
 
-export const updateAddress = async ({
-  addressId,
+export const addAddress = async ({
   formData,
   reCaptchaToken,
 }: {
-  addressId: number;
   formData: FormData;
   reCaptchaToken?: string;
 }) => {
-  const t = await getTranslations('Account.Addresses.UpdateAddress');
+  const t = await getTranslations('Account.Addresses.Add.Form');
 
   const customerId = await getSessionCustomerId();
 
   try {
     const parsed = parseAccountFormData(formData);
 
-    if (!isUpdateCustomerAddressInput(parsed)) {
+    if (!isAddCustomerAddressInput(parsed)) {
       return {
         status: 'error',
         error: t('Errors.inputError'),
@@ -75,19 +70,16 @@ export const updateAddress = async ({
     }
 
     const response = await client.fetch({
-      document: UpdateCustomerAddressMutation,
+      document: AddCustomerAddressMutation,
       customerId,
       fetchOptions: { cache: 'no-store' },
       variables: {
-        input: {
-          addressEntityId: addressId,
-          data: parsed,
-        },
+        input: parsed,
         ...(reCaptchaToken && { reCaptchaV2: { token: reCaptchaToken } }),
       },
     });
 
-    const result = response.data.customer.updateCustomerAddress;
+    const result = response.data.customer.addCustomerAddress;
 
     revalidatePath('/account/addresses', 'page');
 
@@ -97,15 +89,7 @@ export const updateAddress = async ({
 
     return {
       status: 'error',
-      message: result.errors
-        .map((error) => {
-          if (error.__typename === 'AddressDoesNotExistError') {
-            return t('Errors.notFound');
-          }
-
-          return error.message;
-        })
-        .join('\n'),
+      message: result.errors.map((error) => error.message).join('\n'),
     };
   } catch (error: unknown) {
     if (error instanceof Error) {
