+++ core/app/[locale]/(default)/account/(tabs)/addresses/edit/[slug]/_components/edit-address-form.tsx
@@ -12,6 +12,7 @@ import {
   DateField,
   FieldNameToFieldId,
   FieldWrapper,
+  getPreviouslySubmittedValue,
   MultilineText,
   NumbersOnly,
   Password,
@@ -20,16 +21,7 @@ import {
   RadioButtons,
   Text,
 } from '~/components/form-fields';
-import { Link } from '~/components/link';
-import { Button } from '~/components/ui/button';
-import { Field, Form, FormSubmit } from '~/components/ui/form';
-import { Message } from '~/components/ui/message';
-
-import { useAccountStatusContext } from '../../_components/account-status-provider';
-import { addAddress } from '../_actions/add-address';
-
 import {
-  createCountryChangeHandler,
   createDatesValidationHandler,
   createMultilineTextValidationHandler,
   createNumbersInputValidationHandler,
@@ -38,21 +30,54 @@ import {
   createPreSubmitPicklistValidationHandler,
   createRadioButtonsValidationHandler,
   createTextInputValidationHandler,
-} from './address-field-handlers';
-import { NewAddressQueryResponseType } from './customer-new-address';
+  type FieldStateSetFn,
+} from '~/components/form-fields/shared/field-handlers';
+import { Link } from '~/components/link';
+import { Button } from '~/components/ui/button';
+import { Field, Form, FormSubmit } from '~/components/ui/form';
+import { Message } from '~/components/ui/message';
+
+import { useAccountStatusContext } from '../../../../_components/account-status-provider';
+import { Modal } from '../../../../_components/modal';
+import { deleteAddress } from '../../../_actions/delete-address';
+import { updateAddress } from '../_actions/update-address';
+import { CustomerEditAddressQueryResult } from '../page';
 
 interface FormStatus {
   status: 'success' | 'error';
   message: string;
 }
 
-export type AddressFields = NonNullable<
-  NewAddressQueryResponseType['site']['settings']
+type Address = NonNullable<
+  NonNullable<CustomerEditAddressQueryResult['customer']>['addresses']['edges']
+>[number]['node'];
+
+type AddressFields = NonNullable<
+  CustomerEditAddressQueryResult['site']['settings']
 >['formFields']['shippingAddress'];
 
-export type Countries = NonNullable<NewAddressQueryResponseType['geography']['countries']>;
-type CountryCode = Countries[number]['code'];
+type Countries = NonNullable<CustomerEditAddressQueryResult['geography']['countries']>;
 type CountryStates = Countries[number]['statesOrProvinces'];
+
+type FieldUnionType = Exclude<
+  keyof typeof FieldNameToFieldId,
+  'email' | 'password' | 'confirmPassword' | 'exclusiveOffers' | 'currentPassword'
+>;
+
+const createCountryChangeHandler =
+  (provinceSetter: FieldStateSetFn<CountryStates>, countries: Countries) => (value: string) => {
+    const states = countries.find(({ code }) => code === value)?.statesOrProvinces;
+
+    provinceSetter(states ?? []);
+  };
+
+const isExistedField = (name: unknown): name is FieldUnionType => {
+  if (typeof name === 'string' && name in FieldNameToFieldId) {
+    return true;
+  }
+
+  return false;
+};
 
 interface SumbitMessages {
   messages: {
@@ -76,61 +101,62 @@ const SubmitButton = ({ messages }: SumbitMessages) => {
   );
 };
 
-interface AddAddressProps {
+interface EditAddressProps {
+  address: Address;
   addressFields: AddressFields;
   countries: Countries;
-  defaultCountry: {
-    id: number;
-    code: CountryCode;
-    states: CountryStates;
-  };
+  isAddressRemovable: boolean;
   reCaptchaSettings?: {
     isEnabledOnStorefront: boolean;
     siteKey: string;
   };
 }
 
-export const AddAddress = ({
+export const EditAddressForm = ({
+  address,
   addressFields,
   countries,
-  defaultCountry,
+  isAddressRemovable,
   reCaptchaSettings,
-}: AddAddressProps) => {
+}: EditAddressProps) => {
   const form = useRef<HTMLFormElement>(null);
   const [formStatus, setFormStatus] = useState<FormStatus | null>(null);
+  const t = useTranslations('Account.Addresses.Edit.Form');
 
   const reCaptchaRef = useRef<ReCaptcha>(null);
   const router = useRouter();
-  const t = useTranslations('Account.Addresses');
   const [reCaptchaToken, setReCaptchaToken] = useState('');
   const [isReCaptchaValid, setReCaptchaValid] = useState(true);
-
-  const [textInputValid, setTextInputValid] = useState<Record<string, boolean>>({});
-  const [numbersInputValid, setNumbersInputValid] = useState<Record<string, boolean>>({});
-  const [datesValid, setDatesValid] = useState<Record<string, boolean>>({});
-  const [passwordValid, setPasswordValid] = useState<Record<string, boolean>>({});
-  const [radioButtonsValid, setRadioButtonsValid] = useState<Record<string, boolean>>({});
-  const [picklistValid, setPicklistValid] = useState<Record<string, boolean>>({});
-  const [checkboxesValid, setCheckboxesValid] = useState<Record<string, boolean>>({});
-  const [multiTextValid, setMultiTextValid] = useState<Record<string, boolean>>({});
-  const [countryStates, setCountryStates] = useState(defaultCountry.states);
-
   const { setAccountState } = useAccountStatusContext();
 
   useEffect(() => {
     setAccountState({ status: 'idle' });
   }, [setAccountState]);
 
+  const [textInputValid, setTextInputValid] = useState<Record<string, boolean>>({});
+  const [passwordValid, setPasswordValid] = useState<Record<string, boolean>>({});
+  const [numbersInputValid, setNumbersInputValid] = useState<Record<string, boolean>>({});
+  const [datesValid, setDatesValid] = useState<Record<string, boolean>>({});
+  const [radioButtonsValid, setRadioButtonsValid] = useState<Record<string, boolean>>({});
+  const [picklistValid, setPicklistValid] = useState<Record<string, boolean>>({});
+  const [checkboxesValid, setCheckboxesValid] = useState<Record<string, boolean>>({});
+  const [multiTextValid, setMultiTextValid] = useState<Record<string, boolean>>({});
+
+  const defaultStates = countries
+    .filter((country) => country.code === address.countryCode)
+    .flatMap((country) => country.statesOrProvinces);
+  const [countryStates, setCountryStates] = useState<CountryStates>(defaultStates);
+
   const handleTextInputValidation = createTextInputValidationHandler(
     setTextInputValid,
     textInputValid,
   );
   const handlePasswordValidation = createPasswordValidationHandler(setPasswordValid, addressFields);
-  const handleCountryChange = createCountryChangeHandler(setCountryStates, countries);
   const handleNumbersInputValidation = createNumbersInputValidationHandler(
     setNumbersInputValid,
     numbersInputValid,
   );
+  const handleCountryChange = createCountryChangeHandler(setCountryStates, countries);
   const handleDatesValidation = createDatesValidationHandler(setDatesValid, datesValid);
   const handleRadioButtonsChange = createRadioButtonsValidationHandler(
     setRadioButtonsValid,
@@ -140,6 +166,17 @@ export const AddAddress = ({
     setMultiTextValid,
     multiTextValid,
   );
+
+  const onReCaptchaChange = (token: string | null) => {
+    if (!token) {
+      setReCaptchaValid(false);
+
+      return;
+    }
+
+    setReCaptchaToken(token);
+    setReCaptchaValid(true);
+  };
   const validatePicklistFields = createPreSubmitPicklistValidationHandler(
     addressFields,
     setPicklistValid,
@@ -157,16 +194,6 @@ export const AddAddress = ({
     }
   };
 
-  const onReCaptchaChange = (token: string | null) => {
-    if (!token) {
-      setReCaptchaValid(false);
-
-      return;
-    }
-
-    setReCaptchaToken(token);
-    setReCaptchaValid(true);
-  };
   const onSubmit = async (formData: FormData) => {
     if (reCaptchaSettings?.isEnabledOnStorefront && !reCaptchaToken) {
       setReCaptchaValid(false);
@@ -176,12 +203,12 @@ export const AddAddress = ({
 
     setReCaptchaValid(true);
 
-    const submit = await addAddress({ formData, reCaptchaToken });
+    const submit = await updateAddress({ addressId: address.entityId, formData });
 
     if (submit.status === 'success') {
       setAccountState({
         status: 'success',
-        message: t('addNewAddressSuccessMessage'),
+        message: submit.message || '',
       });
 
       router.push('/account/addresses');
@@ -198,6 +225,19 @@ export const AddAddress = ({
       behavior: 'smooth',
     });
   };
+  const onDeleteAddress = async () => {
+    const submit = await deleteAddress(address.entityId);
+
+    if (submit.status === 'success') {
+      setAccountState({ status: submit.status, message: submit.message || '' });
+    }
+
+    if (submit.status === 'error') {
+      setAccountState({ status: submit.status, message: submit.message || '' });
+    }
+
+    router.push('/account/addresses');
+  };
 
   return (
     <>
@@ -211,12 +251,21 @@ export const AddAddress = ({
           {addressFields.map((field) => {
             const fieldId = field.entityId;
             const fieldName = createFieldName(field, 'address');
+            const key = FieldNameToFieldId[fieldId];
+            const defaultCustomField = address.formFields.find(
+              ({ entityId }) => entityId === fieldId,
+            );
+            const defaultValue = (isExistedField(key) && address[key]) || undefined;
 
             switch (field.__typename) {
               case 'TextFormField': {
+                const previousTextValue =
+                  getPreviouslySubmittedValue(defaultCustomField).TextFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <Text
+                      defaultValue={defaultValue ?? previousTextValue}
                       field={field}
                       isValid={textInputValid[fieldId]}
                       name={fieldName}
@@ -227,9 +276,13 @@ export const AddAddress = ({
               }
 
               case 'MultilineTextFormField': {
+                const previousMultiTextValue =
+                  getPreviouslySubmittedValue(defaultCustomField).MultilineTextFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <MultilineText
+                      defaultValue={previousMultiTextValue}
                       field={field}
                       isValid={multiTextValid[fieldId]}
                       name={fieldName}
@@ -240,9 +293,13 @@ export const AddAddress = ({
               }
 
               case 'NumberFormField': {
+                const previousNumberValue =
+                  getPreviouslySubmittedValue(defaultCustomField).NumberFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <NumbersOnly
+                      defaultValue={previousNumberValue}
                       field={field}
                       isValid={numbersInputValid[fieldId]}
                       name={fieldName}
@@ -253,9 +310,13 @@ export const AddAddress = ({
               }
 
               case 'CheckboxesFormField': {
+                const previousCheckboxesValue =
+                  getPreviouslySubmittedValue(defaultCustomField).CheckboxesFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <Checkboxes
+                      defaultValue={previousCheckboxesValue}
                       field={field}
                       isValid={checkboxesValid[fieldId]}
                       name={fieldName}
@@ -267,9 +328,13 @@ export const AddAddress = ({
               }
 
               case 'DateFormField': {
+                const previousDateValue =
+                  getPreviouslySubmittedValue(defaultCustomField).DateFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <DateField
+                      defaultValue={previousDateValue}
                       field={field}
                       isValid={datesValid[fieldId]}
                       name={fieldName}
@@ -281,9 +346,13 @@ export const AddAddress = ({
               }
 
               case 'RadioButtonsFormField': {
+                const previousMultipleChoiceValue =
+                  getPreviouslySubmittedValue(defaultCustomField).MultipleChoiceFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <RadioButtons
+                      defaultValue={previousMultipleChoiceValue}
                       field={field}
                       isValid={radioButtonsValid[fieldId]}
                       name={fieldName}
@@ -295,23 +364,20 @@ export const AddAddress = ({
 
               case 'PicklistFormField': {
                 const isCountrySelector = fieldId === FieldNameToFieldId.countryCode;
+                const previousMultipleChoiceValue =
+                  getPreviouslySubmittedValue(defaultCustomField).MultipleChoiceFormField;
                 const picklistOptions = isCountrySelector
                   ? countries.map(({ name, code }) => ({ label: name, entityId: code }))
                   : field.options;
-                const defaultMultipleChoiceValue = isCountrySelector
-                  ? defaultCountry.code
-                  : undefined;
 
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <Picklist
-                      defaultValue={defaultMultipleChoiceValue}
+                      defaultValue={isCountrySelector ? defaultValue : previousMultipleChoiceValue}
                       field={field}
                       isValid={picklistValid[fieldId]}
                       name={fieldName}
-                      onChange={
-                        fieldId === FieldNameToFieldId.countryCode ? handleCountryChange : undefined
-                      }
+                      onChange={isCountrySelector ? handleCountryChange : undefined}
                       onValidate={setPicklistValid}
                       options={picklistOptions}
                     />
@@ -319,28 +385,30 @@ export const AddAddress = ({
                 );
               }
 
-              case 'PicklistOrTextFormField':
+              case 'PicklistOrTextFormField': {
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <PicklistOrText
                       defaultValue={
-                        fieldId === FieldNameToFieldId.stateOrProvince
-                          ? countryStates[0]?.name
-                          : undefined
+                        fieldId === FieldNameToFieldId.stateOrProvince ? defaultValue : undefined
                       }
                       field={field}
+                      key={countryStates.length}
                       name={fieldName}
-                      options={countryStates.map(({ name }) => {
-                        return { entityId: name, label: name };
-                      })}
+                      options={countryStates.map(({ name }) => ({ entityId: name, label: name }))}
                     />
                   </FieldWrapper>
                 );
+              }
 
               case 'PasswordFormField': {
+                const previousPasswordValue =
+                  getPreviouslySubmittedValue(defaultCustomField).PasswordFormField;
+
                 return (
                   <FieldWrapper fieldId={fieldId} key={fieldId}>
                     <Password
+                      defaultValue={previousPasswordValue}
                       field={field}
                       isValid={passwordValid[fieldId]}
                       name={fieldName}
@@ -371,13 +439,26 @@ export const AddAddress = ({
           )}
         </div>
 
-        <div className="mt-8 flex flex-col justify-stretch gap-2 md:flex-row md:justify-start md:gap-6">
+        <div className="mt-8 flex flex-col justify-stretch gap-2 md:flex-row md:justify-between md:gap-0">
           <FormSubmit asChild>
             <SubmitButton messages={{ submit: t('submit'), submitting: t('submitting') }} />
           </FormSubmit>
-          <Button asChild className="items-center px-8 md:w-fit" variant="secondary">
+          <Button asChild className="items-center px-8 md:ms-6 md:w-fit" variant="secondary">
             <Link href="/account/addresses">{t('cancel')}</Link>
           </Button>
+          <Modal
+            actionHandler={onDeleteAddress}
+            confirmationText={t('confirmDeleteAddress')}
+            title={t('deleteModalTitle')}
+          >
+            <Button
+              className="ms-auto items-center px-8 md:w-fit"
+              disabled={!isAddressRemovable}
+              variant="subtle"
+            >
+              {t('deleteButton')}
+            </Button>
+          </Modal>
         </div>
       </Form>
     </>
