+++ core/components/form-fields/index.ts
@@ -1,12 +1,11 @@
-export * from './date';
 export * from './checkboxes';
-export * from './password';
-export * from './text';
+export * from './date';
 export * from './multiline-text';
 export * from './numbers-only';
-export * from './radio-buttons';
+export * from './password';
 export * from './picklist';
 export * from './picklist-or-text';
+export * from './radio-buttons';
 export * from './shared/field-wrapper';
+export * from './text';
 export * from './utils';
-export * from './parse-fields';
